requirements.txt
chromadb==0.5.23
fastapi==0.115.6
langchain_chroma==0.1.4
langchain_community==0.3.14
langchain_text_splitters==0.3.4
openai==1.59.3
python-dotenv==1.0.1
